requirements.txt
llama-index 
llama-index-vector-stores-mongodb 
llama-index-embeddings-openai 
pymongo
Flask
python-dotenv
llama-index-readers-web
firecrawl-py
google-auth-oauthlib
beautifulsoup4
google-api-python-client
llama-index-postprocessor-cohere-rerank